pyproject.toml
[tool.poetry]
name = "jijenga-referral-system"
version = "0.1.0"
description = "Backend service for the Jijenga Referral System."
authors = ["Your Name <you@example.com>"] # Replace with actual author info
readme = "README.md" # We should create a README later
packages = [{include = "app"}] # Include the 'app' directory as a package

[tool.poetry.dependencies]
python = "^3.10" # Or the appropriate Python version
fastapi = "^0.110.0" # Use compatible versions
sqlalchemy = "^2.0.0"
alembic = "^1.13.0"
asyncpg = "^0.29.0"
pydantic = {extras = ["email"], version = "^2.11.7"}
pydantic-settings = "^2.2.0"
passlib = {extras = ["bcrypt"], version = "^1.7.0"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
arq = "^0.25.0"
httpx = "^0.27.0"
uvicorn = {extras = ["standard"], version = "^0.29.0"} # Add uvicorn for running the app
psycopg2 = "^2.9.10"

[tool.poetry.group.dev.dependencies]
pytest = "^8.1.0"
pytest-asyncio = "^0.23.0"
pytest-mock = "^3.12.0"
ipython = "^8.22.0" # Optional: useful for development
pytest-dotenv = "^0.5.2" # Add pytest-dotenv

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = "."
